backend/requirements.txt
Django==5.1
gunicorn==25.1.0
paho-mqtt==2.1.0
psycopg2-binary==2.9.11
python-dotenv==1.2.2
reportlab==4.4.4
